fix: use opencl3 instead of ocl
Also refactor the code. Instead of operating on the `Buffer`s, operate on
the `Program` and pass in the `Buffer`s. This way there is no need to expose
the queue to the outside.

Kernel creation might now return an error if the kernel with the given
name doesn't exist.