refactor: devices are now a vector
Prior to this change the devices were a hash map where the key is the
brand and the value are the corresponding devices. This was an
unnecessary indirection as we only care about the devices. The devices
still contain the brand information, so you can filter a certain
brand by looping through the devices.

This commit also makes `DEVICES` only public to the crate and removes
the global public `PLATFORMS`.

These code change won't break the current upstream users bellperson and
neptune.