feat: better unique identifier support
Using the UUID as unique identifier is preferred, though it is not always
available. Hence there is a fallback to the PCI ID, which is almost always
available on all supported devices. The only case it is not available is
on MacBooks with an AMD graphics card. Those machines are used for
development, but not in production.

BREAKING CHANGE: Bus ID is replaced with PCI ID

Prior to this change only the so-called "Bus ID" was used. Instead now use
a combination of the device and the Bus Id, which should provide more
uniqueness.

Also the `GPUSelector` is replaced by with additional functions on
`Device` which provide the same functionality.