fix: identify AMD cards correctly
There were problems identifying AMD cards correctly [1]. The reason is
that the code so far was tested with integrated AMD GPUs on Apple
devices, which report "AMD" as vendor name. Though usually the vendor
name of AMD is "Advanced Micro Devices, Inc.". Both should be supported.

As the code needed to change anyway, I decided to switch from string
identifiers to using the vendor ID, which is a numeric value. This
should lead to a more robust identification.

This commit also changes the `Display` string for AMD, it's now using
the more common "Advanced Micro Devices, Inc." instead of "AMD".

[1]: https://github.com/filecoin-project/rust-gpu-tools/issues/45