feat: add core counts
For maxing out a GPU it's sometimes useful to know the number of CUDA cores.
This commit adds a mapping between graphic cards name and their number of
CUDA cores.

For non CUDA cards an appropriate number based on the compute units is chosen.

If your card isn't specified, you can add it to the list via the
`RUST_GPU_TOOLS_CUSTOM_CPU` environment variable. Example:

    RUST_GPU_TOOLS_CUSTOM_GPU="GeForce RTX 2080 Ti:4352,GeForce GTX 1060:1280"

This code was originally developed as part of bellperson:
https://github.com/filecoin-project/bellperson/blob/dcf8a467a05796cf0f7fb9e170b6fc6845b42d9f/src/gpu/utils.rs#L6-L56