feat: replace RustaCUDA with cust
RustaCUDA doesn't seem to maintained anymore. Instead of maintaining a fork
of it, switch to cust. For now use the legacy context instead of the new
primary context.

BREAKING CHANGE: `cuda::from_binary` was renamed to `from_file`

It also takes a `Path` as input instead of a `CStr` now.