fix: remove the core counts (#77)
The core counts are not really meaningful, what matters are the
compute units/streaming-multiprocessor units. Thefore access to
those numbers was introduced in
https://github.com/filecoin-project/rust-gpu-tools/commit/775e63327e6836e90afd458f16d6f0707345600a
which removes the need for the core counts.

This commit removes the notion of core counts from this libarary.

BREAKING CHANGE: The `CUDA_CORES` mapping was removed.
Use the `compute_units` proerty of the `Device` instead.