feat: switch to once_cell (#81)
Prior to this change, we were using `lazy_static`. While this worked, there
is a problem if the `lazy_static` `spin_no_std` features gets enabled. It
then requires the types used to implement `Send`.

`once_cell` has the same requirement, which was one of the reasons not to
use it. Though it makes sense to implement `Send` for the `CudaContexts`.
This means that now the more ergonomic `once_cell` can be used.

Fixes #66.